### 2022-07-28 #4614 @pydemull
Hello,
Potential reviewers could be the followings (based on their preferences regarding R and Shiny):

  *   Ecosantos/
  *   PhilipPallmann
  *   EDiLD
  *   sandrinecharles
  *   luciorq
  *   elimillera
  *   adithirgis
  *   aj2duncan

Sincerely,
Pierre-Yves de Müllenheim

________________________________
De : The Open Journals editorial robot ***@***.***>
Envoyé : mercredi 27 juillet 2022 13:50
À : openjournals/joss-reviews ***@***.***>
Cc : Pierre-Yves de Müllenheim ***@***.***>; Mention ***@***.***>
Objet : [openjournals/joss-reviews] [PRE REVIEW]: activAnalyzer: An R Shiny app to analyse ActiGraph accelerometer data and to implement the use of the PROactive Physical Activity in COPD instruments (Issue #4614)

Submitting author: @pydemull<https://github.com/pydemull> (Pierre-Yves de Müllenheim<http://orcid.org/0000-0001-9157-7371>)
Repository: https://github.com/pydemull/activAnalyzer
Branch with paper.md (empty if default branch):
Version: v1.0.4
Editor: Pending
Reviewers: Pending
Managing EiC: Kevin M. Moerman

Status

[status]<https://joss.theoj.org/papers/5d6659af1bf8ca2fb977c189039b8315>

Status badge code:

Author instructions

Thanks for submitting your paper to JOSS @pydemull<https://github.com/pydemull>. Currently, there isn't a JOSS editor assigned to your paper.

@pydemull<https://github.com/pydemull> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people<https://bit.ly/joss-reviewers> have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Editor instructions

The JOSS submission bot @editorialbot<https://github.com/editorialbot> is here to help you find and assign reviewers and start the main review. To find out what @editorialbot<https://github.com/editorialbot> can do for you type:
