+++ README.Rmd
@@ -18,6 +18,7 @@ knitr::opts_chunk$set(
 [![Codecov test coverage](https://codecov.io/gh/pydemull/activAnalyzer/branch/master/graph/badge.svg)](https://app.codecov.io/gh/pydemull/activAnalyzer?branch=master)
 [![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
 [![R-CMD-check](https://github.com/pydemull/activAnalyzer/workflows/R-CMD-check/badge.svg)](https://github.com/pydemull/activAnalyzer/actions)
+[![](http://cranlogs.r-pkg.org/badges/grand-total/activAnalyzer?color=blue)](https://cran.r-project.org/package=activAnalyzer)
 <!-- badges: end -->
 
 activAnalyzer is a Shiny app that has been developed to analyze daily physical behavior data recorded at the hip in adults using an ActiGraph accelerometer (.agd files). Once analysis is completed, the app allows exporting results (summarized by day with means or medians of the valid days) to .csv files and generating a report of the measurement (.pdf file). All the configured inputs relevant for interpreting the results are recorded in the report. Be sure that the inputs that are configured when generating the report correspond to the analysis that was actually performed (in other words, avoid modifying the inputs after generating satisfactory results). In addition to a general analysis of physical behavior, the app also allows to implement the [Daily- and Clinical visit-PROactive Physical Activity in COPD (chronic obstructive pulmonary disease) instruments (D-PPAC and C-PPAC)](https://erj.ersjournals.com/content/early/2015/05/28/09031936.00183014). Please read the [user's guide](https://github.com/pydemull/activAnalyzer/blob/master/inst/guide/user_guide_en.pdf) for details about how the app works.
