### 2022-08-12 #4614 @arfon
@samhforbes – you can just say `@editorialbot assign me as editor`
