### 2022-08-12 #4614 @arfon
Yep, invites should work now too. I just needed to add you to the `joss-eics` GitHub team.
