### 2022-08-17 #4614 @samhforbes
Thanks for submitting to JOSS @pydemull. I have reached out to some prospective reviewers and they will be added here as they accept the invitation. I’m looking forward to working through this process with you.
