### 2022-08-30 #4614 @samhforbes
Hi @ellessenne would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Please note this is a public issue, and your responses will also be public.
