### 2022-08-30 #4614 @sandrinecharles
Dear @samhforbes I am sorry but I am not available these next five weeks due to the beginning of the beginning of the university year. Since I am responsible for the Doctoral School (250 PhD students in total), I do not have a lot of time for reviewing. I suggest you to contact either Virgile BAUDROT (virgile.baudrot@qonfluens.com) and/or Aude RATIER (aude.ratier@ineris.fr). Best regards, Sandrine
