### 2022-09-05 #4741 @editorialbot
[REVIEW]: activAnalyzer: An R Shiny app to analyse ActiGraph accelerometer data and to implement the use of the PROactive Physical Activity in COPD instruments
**Submitting author:** @pydemull (<a href="http://orcid.org/0000-0001-9157-7371">Pierre-Yves de Müllenheim</a>)
**Repository:** https://github.com/pydemull/activAnalyzer
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
