### 2022-09-30 pydemull/activAnalyzer#10 @angerhang
Limited gender category
The current `gender` field has the following types:
* male
* female
* undefined

Ideally these types should be more inclusive. Non-binary individuals are best not phrased at undefined. This [post](https://uxdesign.cc/designing-forms-for-gender-diversity-and-inclusion-d8194cf1f51) contains some nice suggestions.

![Screenshot 2022-09-30 at 23 40 11](https://user-images.githubusercontent.com/7655454/193365705-6a54c5b7-9f90-4a8f-9c1b-5b8105363288.png)
