### 2022-09-30 pydemull/activAnalyzer#11 @angerhang
Good to have additional dependency specifications
It would be good to know the key dependencies this package requires, for example, tested R and R studio version. It will make it easier for others to debug/contribute in the future.
