### 2022-09-30 #4741 @angerhang
Overall, I found `activAnalyzer` easy to navigate and use. The documentation and paper are well-written and easy to follow. So happy to see another open source software trying to reduce the hurdle of the costly ActiLife software in the community. The added integration of COPD questionnaires is excellent. I am happy to tick most of the checkboxes above with a few minor comments:

* [ Limited gender category in the survey UI](https://github.com/pydemull/activAnalyzer/issues/10)
* [ Good to clarify high-level dependencies for both debugging and others to contribute in the future](https://github.com/pydemull/activAnalyzer/issues/11)
* It would also be good to provide a sample `.agd` file for us so that can verify that the package works? It might also be a good idea to include this test file in your CI tests to ensure that your data gets processed properly.

@pydemull thanks a lot for making this valuable contribution to the community. I look forward to your revisions and seeing the generated report myself using some of the test files before signing off the rest of the checklist.

Great job, nontheless :D
