+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   activAnalyzer: activAnalyzer.html
-last_built: 2022-08-04T21:01Z
+last_built: 2022-10-05T16:04Z
 urls:
   reference: https://pydemull.github.io/activAnalyzer/reference
   article: https://pydemull.github.io/activAnalyzer/articles
