+++ DESCRIPTION
@@ -13,17 +13,17 @@ License: GPL (>= 3)
 Imports: 
     dbplyr (>= 2.1.1),
     dplyr (>= 1.0.10),
-    flextable (>= 0.7.0),
-    forcats (>= 0.5.1),
+    flextable (>= 0.8.2),
+    forcats (>= 0.5.2),
     ggplot2 (>= 3.3.6),
-    golem (>= 0.3.2),
+    golem (>= 0.3.4),
     hms (>= 1.1.2),
     lubridate (>= 1.8.0),
-    magrittr,
-    modelr (>= 0.1.8),
-    patchwork (>= 1.1.1),
+    magrittr (>= 2.0.3),
+    modelr (>= 0.1.9),
+    patchwork (>= 1.1.2),
     PhysicalActivity,
-    reactable (>= 0.2.3),
+    reactable (>= 0.3.0),
     rmarkdown (>= 2.16),
     RSQLite,
     shiny (>= 1.7.2),
@@ -38,7 +38,7 @@ Imports:
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.2
+RoxygenNote: 7.2.1
 URL: https://pydemull.github.io/activAnalyzer/, https://github.com/pydemull/activAnalyzer
 BugReports: https://github.com/pydemull/activAnalyzer/issues
 Suggests: 
