### 2022-10-05 c00b800
maj function documentation (cppac/dppac)
