### 2022-10-05 e9dfb96
added arguments to icon() functions in the UI to avoid error messages when running app
