### 2022-10-05 bd0e433
added in the README the minimum R and RStudio versions recommended for using the app
