### 2022-10-07 pydemull/activAnalyzer#11 @pydemull
Thank you @angerhang for your comment. I have added the followings into the revised version of the README (Usage section, 3rd bullet point, bold text below):

"On your machine via [R software](https://cran.r-project.org/) **(version: ≥ 3.4.0)**, the [RStudio environment](https://www.rstudio.com/) **(recommended versions: Prairie Trillium [2022.02] and newer)**, and the [activAnalyzer package](https://github.com/pydemull/activAnalyzer) installable from CRAN (stable version) or from GitHub (development version)."

The minimum R version to be used is also indicated in the DESCRIPTION file.

Please let me know if it is not enough.
