### 2022-10-07 4525c0d
updated mark_intensity() function: correction for the numbers provided in the intensity_category_num variable of the returned dataset.
