### 2022-10-12 #4741 @pydemull
Dear @samhforbes @angerhang @elimillera, CRAN has requested I make some updates of the package so that it passes R-dev checking (this is in relation with a planned update of R regarding as.character.POSIXt()). I have made the update and the current CRAN version is now 1.0.5. (Thus, I will have to make a change in the JOSS paper where I gave a version number for the package; I will remove the version number for the package).
Sincerely
