### 2022-10-16 #4741 @angerhang
> > Thank you very much @angerhang for your comments. Before making the required changes to the package, I have some questions to be sure of what you want.
> >
> > 1. Regarding the issue related to the "gender" options: Actually, it was not "Gender" (that refers to a "constellation of sociocultural processes that interact with and have the potential to influence human biology" [Schiebinger, et al. Lancet. 2016]) but "Sex" (that refers to "a biological variable based upon chromosomal assignment, and generally male, female, or intersex" [Schiebinger, op. cit.]) that should be provided. While I totally respect and am happy with all the gender categories, there are two reasons for which I choose sex information only. The first one is that the equations used in the app (and that have been previously published) actually use a binary classification (male/female, men/women) and I suspect that the studies that used the "gender" term when they did not use the "sex" term, actually took (reported) sex information, because there are of course more than only two gender categories. The second reason is that gender may be more complex to assess, at least when reported (but I may be wrong when saying this), and because the app did not need this information, I prefered to avoid this difficulty. However, in the editorial by Schiebinger cited above, the term "intersex" is evocated, and this may be an information confirming that my options are indeed incomplete. Would it be ok for you if I keep the "sex" information with the following options: "male", "female", "intersex", "undefined"?
> > 2. Regarding your remarks for the .agd file: Actually, once the package is installed locally, you can access the file by the following code (as shown in the examples of the functions): file <- system.file("extdata", "acc.agd", package = "activAnalyzer"). This is the data file that is loaded when the user clicks on the dedicated button to choose the "demo file" inside the app. This also is this file that is used for my CI tests that are performed thanks to the files that can be found in the "tests" folder of the package. So yes, there are CI tests (or maybe you though about other things?). Based on this comment, would you prefer I make available the agd file content when the user uses the data() function? Or my explanations are now sufficient for you?
> >
> > Thanks a lot for your time and work.
>
> 1. Thanks a lot for your response @pydemull . Now I understand the rationale of your design choice. Thanks for being willing to build a more inclusive scientific software. Perhaps `intersex` and `undefined` will work? I guess with non-binary individuals, you will just take an average of both sex for your calculation?
> 2. I think your explanations are sufficient. I will try testing the software using the sample data this weekend. Thanks a lot!

Tested with the demo data. The report generated is super sick! Am sure many will benefit from using this software. There are still outstanding tasks regarding the sex calculation and dependency instructions. But I am happy to tick all the boxes from my checklist :D
