### 2022-10-21 0a338a1
added "intersex" and "prefer not to say" categories for sex classification; updated the `compute_bmr()`and `compute_mets()` functions that now use the `dplyr::case_when()` function to provide results (rather than `ifelse()` function).
