+++ tests/testthat/test-compute_mets.R
@@ -1,84 +1,120 @@
 
 test_that("METs are correctly computed with Santos-Lozano et al. equation [older adults]", {
   
-  MET_santos_older_man_male <- 2.5878 + 0.00047 * 1500 - 0.6453 * 2 # computation for a male
-  MET_santos_older_man_female <- 2.5878 + 0.00047 * 1500 - 0.6453 * 1 # computation for a female
-  MET_santos_older_man_undefined <- 2.5878 + 0.00047 * 1500 - 0.6453 * 1 # computation for a undefined sex
+  MET_santos_older_male <- 2.5878 + 0.00047 * 1500 - 0.6453 * 2 # computation for a male
+  MET_santos_older_female <- 2.5878 + 0.00047 * 1500 - 0.6453 * 1 # computation for a female
+  MET_santos_older_undefined <- 2.5878 + 0.00047 * 1500 - 0.6453 * 1 # computation for an undefined sex
+  MET_santos_older_intersex <- mean(c(MET_santos_older_male, MET_santos_older_female)) # computation for intersex
   
-  MET_santos_older_func_male <- compute_mets(data = data.frame(vm = 1500), 
+
+  MET_santos_older_func_male <- compute_mets(data = data.frame(axis1 = 1995, vm = 1500), 
                                         equation = "Santos-Lozano et al. (2013) [Older adults]", sex = "male")
-  MET_santos_older_func_female <- compute_mets(data = data.frame(vm = 1500), 
+  MET_santos_older_func_female <- compute_mets(data = data.frame(axis1 = 1995, vm = 1500), 
                                              equation = "Santos-Lozano et al. (2013) [Older adults]", sex = "female")
-  MET_santos_older_func_undefined <- compute_mets(data = data.frame(vm = 1500), 
+  MET_santos_older_func_undefined <- compute_mets(data = data.frame(axis1 = 1995, vm = 1500), 
                                                equation = "Santos-Lozano et al. (2013) [Older adults]", sex = "undefined")
-    
-  expect_equal(MET_santos_older_func_male, MET_santos_older_man_male)
-  expect_equal(MET_santos_older_func_female, MET_santos_older_man_female)
-  expect_equal(MET_santos_older_func_undefined, MET_santos_older_man_undefined)
+  MET_santos_older_func_intersex <- compute_mets(data = data.frame(axis1 = 1995, vm = 1500), 
+                                                  equation = "Santos-Lozano et al. (2013) [Older adults]", sex = "intersex")
+  MET_santos_older_func_no_answer <- compute_mets(data = data.frame(axis1 = 1995, vm = 1500), 
+                                                 equation = "Santos-Lozano et al. (2013) [Older adults]", sex = "prefer not to say")
+  
+  
+  expect_equal(MET_santos_older_func_male, MET_santos_older_male)
+  expect_equal(MET_santos_older_func_female, MET_santos_older_female)
+  expect_equal(MET_santos_older_func_undefined, MET_santos_older_undefined)
+  expect_equal(MET_santos_older_func_intersex, MET_santos_older_intersex)
+  expect_equal(MET_santos_older_func_no_answer, MET_santos_older_female)
   
 })
 
 
 test_that("METs are correctly computed with Santos-Lozano et al. equation [Adults]", {
   
-  MET_santos_adult_man_male <- 2.8323 + 0.00054 * 1500 - 0.05912 * 67 + 1.4410 * 2  # computation for a male
-  MET_santos_adult_man_female <- 2.8323 + 0.00054 * 1500 - 0.05912 * 67 + 1.4410 * 1  # computation for a female
[…]
