+++ R/compute_bmr.R
@@ -14,28 +14,39 @@
 #' @examples
 #' compute_bmr(age = 32, sex = "male", weight = 67)
 #' 
-compute_bmr <- function(age = 40, sex = c("male", "female", "undefined"), weight = 70) {
+compute_bmr <- function(age = 40, sex = c("male", "female", "intersex", "undefined", "prefer not to say"), weight = 70) {
   
 sex <- match.arg(sex)
   
 
 # BMR using sex and weight only (Henry, 2005; doi: 10.1079/PHN2005801)
-bmr <-
-  ifelse(age <3 & sex == "male", 61.0 * weight - 33.7,
-         ifelse(age >=3 & age <10 & sex == "male", 23.3 * weight + 514,
-                ifelse(age >=10 & age <18 & sex == "male", 18.4 * weight + 581,
-                       ifelse(age >=18 & age <30 & sex == "male", 16.0 * weight + 545,
-                              ifelse(age >=30 & age <60 & sex == "male", 14.2  * weight + 593,
-                                     ifelse(age >=60 & age <70 & sex == "male", 13.0 * weight + 567,
-                                            ifelse(age >=70 & sex == "male", 13.7 * weight + 481,
+bmr <- dplyr::case_when(
+
+  age <3   & sex == "male"             ~ 61.0 * weight - 33.7,
+  age >=3  & age <10 & sex == "male"   ~ 23.3 * weight + 514,
+  age >=10 & age <18 & sex == "male"   ~ 18.4 * weight + 581,
+  age >=18 & age <30 & sex == "male"   ~ 16.0 * weight + 545,
+  age >=30 & age <60 & sex == "male"   ~ 14.2  * weight + 593,
+  age >=60 & age <70 & sex == "male"   ~ 13.0 * weight + 567,
+  age >=70 & sex == "male"             ~ 13.7 * weight + 481,
                                                    
-                                                   ifelse(age <3 & sex == "female", 58.9 * weight - 23.1,
-                                                          ifelse(age >=3 & age <10 & (sex == "female" | sex == "undefined"), 20.1 * weight + 507,
-                                                                 ifelse(age >=10 & age <18 & (sex == "female" | sex == "undefined"), 11.1 * weight + 761,
-                                                                        ifelse(age >=18 & age <30 & (sex == "female" | sex == "undefined"), 13.1 * weight + 558,
-                                                                               ifelse(age >=30 & age <60 & (sex == "female" | sex == "undefined"), 9.74 * weight + 694,
-                                                                                      ifelse(age >=60 & age <70 & (sex == "female" | sex == "undefined"), 10.2 * weight +  572,
-                                                                                             ifelse(age >=70 & (sex == "female" | sex == "undefined"), 10.0 * weight + 577))))))))))))))
+  age <3   & sex == "female"                                   ~ 58.9 * weight - 23.1,
+  age >=3  & age <10 & (sex == "female" | sex == "undefined" | sex == "prefer not to say")  ~ 20.1 * weight + 507,
+  age >=10 & age <18 & (sex == "female" | sex == "undefined" | sex == "prefer not to say")  ~ 11.1 * weight + 761,
+  age >=18 & age <30 & (sex == "female" | sex == "undefined" | sex == "prefer not to say")  ~ 13.1 * weight + 558,
+  age >=30 & age <60 & (sex == "female" | sex == "undefined" | sex == "prefer not to say")  ~ 9.74 * weight + 694,
[…]
