+++ man/mark_intensity.Rd
@@ -14,7 +14,7 @@ mark_intensity(
     "Freedson et al. (1998) [Adults]", "Santos-Lozano et al. (2013) [Older adults]"),
   age = 40,
   weight = 70,
-  sex = c("male", "female", "undefined"),
+  sex = c("male", "female", "intersex", "undefined", "prefer not to say"),
   dates = NULL
 )
 }

[…]
