### 2022-10-26 #4741 @elimillera
Hey @pydemull, I've reviewed the package, repository, and paper. I've ran all the tests I generally run to determine package quality and they are all coming back green, they repository is well formed and has good CI and contributor information, and the paper is well written.

Looks good to me!
