### 2022-10-27 #4741 @pydemull
Thanks @elimillera for your time and work. Glad to know that all seem ok for you. @samhforbes, @angerhang, for information I have merged the dev branch with the master branch in relation to the modifications made following the @angerhang's comments. I have then taken the liberty to close the issues opened by @angerhang. Best
