### 2022-11-06 1b49406
updated recap_by_day() function (inclusion of the new sex categories that were forgotten in the previous commit)
