### 2022-11-11 08561d0
widjets icons (sidebard) adjustements
