+++ inst/guide/user_guide_en.Rmd
@@ -35,7 +35,7 @@ It is possible to provide the name and surname related to the patient, as well a
 It is possible to indicate where the device was placed on the body during the measurement period (hip or back only). Device information must be provided to have a complete document after generating the report at the end of the app. Other relevant information regarding the device (i.e., ActiGraph model, sampling rate, filter enabled when the .agd file was generated from .gt3x data with ActiLife® software) are silently captured when uploading the data file.  
 
 # Data uploading, nonwear time detection, and data visualization
-The user must upload an .agd file previously generated using ActiLife<span>&#174;</span> software with at least the data related to the three axes, the step counts, and the inclinometer function of the device  (time spent in standing, sitting, lying postures, and off; .agd files obtained with "thigh" set as wearing position in ActiLife<span>&#174;</span> will cannot be analyzed). Then, the user can choose the epoch to be used for analysis. The default is set to 60 s as it is the most commonly used epoch in adults; shorter epochs will slow the subsequent analyses. After this step, it is possible to configure the analysis to be performed to detect nonwear time. It consists of choosing the activity data (vector magnitude counts or vertical axis counts) and the time interval with zero count to be considered to detect nonwear time, as well as the time interval with nonzero counts allowed during a nonwear period along with the period duration with zero count required back and forward the detected activity to validate nonwear time. The default values provided in the app for configuring nonwear time detection are based on the paper by Choi et al. [-@choiAssessmentWearNonwear2012]. Finally, when all inputs are configured as required, the user must click on the "Validate configuration" button. The app will automatically compute the vector magnitude ($VM = \sqrt{x^2 + y^2 + z^2}$) before analyzing nonwear time. If all inputs are valid, the app detects nonwear time thanks to a function from the `PhysicalActivity` R package [@choiPhysicalActivityProcessAccelerometer2021]. The app then provides a graphic allowing the user to visualize different data among those contained in the data file. Completing this step is required before going further in the app.
+The user must upload an .agd file previously generated using ActiLife<span>&#174;</span> software with at least the data related to the three axes and the step counts. Then, the user can choose the epoch to be used for analysis. The default is set to 60 s as it is the most commonly used epoch in adults; shorter epochs will slow the subsequent analyses. After this step, it is possible to configure the analysis to be performed to detect nonwear time. It consists of choosing the activity data (vector magnitude counts or vertical axis counts) and the time interval with zero count to be considered to detect nonwear time, as well as the time interval with nonzero counts allowed during a nonwear period along with the period duration with zero count required back and forward the detected activity to validate nonwear time. The default values provided in the app for configuring nonwear time detection are based on the paper by Choi et al. [-@choiAssessmentWearNonwear2012]. Finally, when all inputs are configured as required, the user must click on the "Validate configuration" button. The app will automatically compute the vector magnitude ($VM = \sqrt{x^2 + y^2 + z^2}$) before analyzing nonwear time. If all inputs are valid, the app detects nonwear time thanks to a function from the `PhysicalActivity` R package [@choiPhysicalActivityProcessAccelerometer2021]. The app then provides a graphic allowing the user to visualize different data among those contained in the data file. Completing this step is required before going further in the app.
 
 # Configuration for metrics computation
 It is possible to select the days with which analyzis has to be performed. Then, the user must select an equation to compute METs and the axis and cut-points to be used to compute time spent in sedentary behavior (SED), light physical activity (LPA), moderate physical activity (MPA), vigorous physical activity (VPA), and moderate-to-vigorous physical activity (MVPA). 
