### 2022-11-12 01b2ac1
changed the `size` arguments of the internal `geom_line()` functions by `linewidth` arguments in relation to the v3.4.0 `{ggplot2}` update.
