### 2022-11-12 60e7f51
added the possibility to zoom in on the figures of the app
