+++ R/app_server.R
@@ -355,7 +355,7 @@ app_server <- function(input, output, session) {
       updateNumericInput(inputId = "allowanceFrame_size", value = 2)
       updateNumericInput(inputId = "streamFrame_size", value = 30)
       updateSelectInput(inputId = "start_day_analysis", selected = hms::as_hms(0))
-      updateSelectInput(inputId = "end_day_analysis", selected = hms::as_hms(60*60*23+60*59))
+      updateSelectInput(inputId = "end_day_analysis", selected = hms::as_hms(60*60*23+60*59+59))
     })
     
   
@@ -363,7 +363,51 @@ app_server <- function(input, output, session) {
   # Visualizing all data with nonwear time ----
   ########################################
   
+  # Creating/updating reactive values for zooming in on the plot with nonwear time
+    zoom_param <- reactiveValues(
+      metric = "axis1", 
+      zoom_from_weartime = "00:00:00", 
+      zoom_to_weartime = "23:59:59"
+    )
+    
+    observeEvent(input$validate, {
+      zoom_param$metric <- "axis1"
+      zoom_param$zoom_from_weartime <- "00:00:00"
+      zoom_param$zoom_to_weartime <- "23:59:59"
+      updateSelectInput(inputId = "Metric", selected = "axis1")
+      updateSelectInput(inputId = "zoom_from_weartime", selected = hms::as_hms(0))
+      updateSelectInput(inputId = "zoom_to_weartime", selected = hms::as_hms(60*60*23+60*59+59))
+    })
+    
+    observeEvent(input$update_graphic, {
+      zoom_param$metric <- input$Metric
+      zoom_param$zoom_from_weartime <- input$zoom_from_weartime
+      zoom_param$zoom_to_weartime <- input$zoom_to_weartime
+    })
+    
+    # Controlling for correct inputs when updating the plot with nonwear time marks
+    observeEvent(input$update_graphic,
+                 shinyFeedback::feedbackWarning(
+                   "zoom_from_weartime", 
+                   hms::as_hms(input$zoom_to_weartime) <= hms::as_hms(input$zoom_from_weartime),
+                   "Start time should be inferior to end time."
+                 )
+    )
+    
+    observeEvent(input$update_graphic,
+                 shinyFeedback::feedbackWarning(
+                   "zoom_to_weartime", 
+                   hms::as_hms(input$zoom_to_weartime) <= hms::as_hms(input$zoom_from_weartime),
+                   "End time should be superior to start time."
+                 )
+    )
+  
   output$graph <- renderPlot({
+    
+    # Waiting for correct inputs
+      req(zoom_param$zoom_from_weartime < zoom_param$zoom_to_weartime)
+    
+    # Making the plot
     
     if (as.numeric(df()$time[2] - df()$time[1]) < 10) { 
       ggplot2::ggplot() + ggplot2::geom_text(
@@ -379,7 +423,12 @@ app_server <- function(input, output, session) {
           axis.ticks = ggplot2::element_blank()
         )
     } else {
-    plot_data(data = df(), metric = input$Metric)
+    plot_data(
+      data = df(), 
+      metric = zoom_param$metric, 
+      zoom_from = zoom_param$zoom_from_weartime,
+      zoom_to = zoom_param$zoom_to_weartime
+      )
     }
   }, 
   width = "auto", 
@@ -693,7 +742,7 @@ app_server <- function(input, output, session) {
                      shinyFeedback::feedbackWarning(
                        "start_day_analysis", 
                        hms::as_hms(input$end_day_analysis) <= hms::as_hms(input$start_day_analysis),
-                       "End time should be superior to start time."
[…]
