### 2022-11-13 48c0728
added tests for graphic updates
