### 2022-11-20 771ca35
add UI elements to indicate relevant missing PA periods
