### 2022-11-21 #4741 @samhforbes
Hi @pydemull this is a neat piece of work, and I like the implementation of the GUI as well.

Please note we have removed a reviewer since they have been too busy to continue the review.
I am looking through the paper and note some small errors. I flag the ones I see here, but perhaps you wouldn't mind having a careful read through as well.

 G-force data are developping -> developing
Such a framework requires to combine scores related to answers to questionnaire items and scores related to accelerometer metrics obtained from a week of measurement. -> Such a framework requires combining scores relating to answers of questionnaire items and scores relating to...

The app can be used according to three different ways -> The app can be used in three different ways
the app will require to install -> the app will require the installation of

he/she has to deal with four ordered sections -> they have to... (and again for the he/she below)
