+++ R/mod_report_pa_period.R
@@ -15,7 +15,7 @@ mod_report_pa_period_ui <- function(id){
              fluidRow(
                
                # Date
-               column(3,
+               column(2,
                       selectInput(
                         ns("corr_date"), 
                         "Date (YYYY-MM-DD)",
@@ -24,28 +24,54 @@ mod_report_pa_period_ui <- function(id){
                ),
                
                # Start time
-               column(3,
-                      selectizeInput(
-                        ns("corr_start_time"),  
-                        "Start time (hh:mm:ss)",
-                        choices = "..."
+               column(1,
+                      numericInput(
+                        ns("corr_start_time_hh"), 
+                        "Start time (hh:)",
+                        value = 0,
+                        min = 0,
+                        max = 23,
+                        step = 1
+                      )
+               ),
+               column(1,
+                      numericInput(
+                        ns("corr_start_time_mm"), 
+                        "Start time (mm:)",
+                        value = 0,
+                        min = 0,
+                        max = 59,
+                        step = 1
                       )
                ),
                
                # End time
-               column(3,
-                      selectizeInput(
-                        ns("corr_end_time"), 
-                        "End time (hh:mm:ss)",
-                        choices = "..."
+               column(1,
+                      numericInput(
+                        ns("corr_end_time_hh"), 
+                        "End time (hh:)",
+                        value = 0,
+                        min = 0,
+                        max = 23,
+                        step = 1
+                      )
+               ),
+               column(1,
+                      numericInput(
+                        ns("corr_end_time_mm"), 
+                        "End time (mm:)",
+                        value = 0,
+                        min = 0,
+                        max = 59,
+                        step = 1
                       )
                ),
                
                # METs
-               column(3,
+               column(1,
                       numericInput(
                         ns("corr_mets"), 
-                        "Activity METs",
+                        "Activity METs\n",
                         value = 0,
                         min = 0,
                         step = 0.1
@@ -60,20 +86,23 @@ mod_report_pa_period_ui <- function(id){
 #' report_pa_period Server Functions
 #'
 #' @noRd 
-mod_report_pa_period_server <- function(id, add_period_btn = NULL, remove_period_btn = NULL){
+mod_report_pa_period_server <- function(id, add_period_btn = NULL, remove_period_btn = NULL, dates_inputs = NULL){
   moduleServer(id, function(input, output, session){
     
     ns <- session$ns
     
-    # Initializing PA periods inputs
-    updateSelectizeInput(session, "corr_start_time", choices = c(hms::as_hms(seq(0, 60*(24*60-1), 60))), server = TRUE)
-    updateSelectizeInput(session, "corr_end_time", choices = c(hms::as_hms(seq(0, 60*(24*60-1), 60))), server = TRUE)
-    
+    # Updating date inputs
+   observe({
+     updateSelectInput(session, "corr_date", choices = c("...", dates_inputs$dates))
+     })
+
     # Hidding inputs when starting the app except the first row of inputs
     if (id != "period_1") {
       shinyjs::hide("corr_date")
-      shinyjs::hide("corr_start_time")
-      shinyjs::hide("corr_end_time")
+      shinyjs::hide("corr_start_time_hh")
+      shinyjs::hide("corr_start_time_mm")
+      shinyjs::hide("corr_end_time_hh")
+      shinyjs::hide("corr_end_time_mm")
       shinyjs::hide("corr_mets")
     }
     
[…]
