+++ R/app_server.R
@@ -451,6 +451,14 @@ app_server <- function(input, output, session) {
   # Adding missing physical activity information if any ----
   #####################################################
   
+  # Initializing reactive values for measurement dates
+  
+  dates_inputs <- reactiveValues(dates = NULL)
+  
+  observeEvent(input$validate, {
+    dates_inputs$dates <- attributes(as.factor(df()$date))$levels
+  })
+  
   # Setting reactive buttons
     period_buttons_1 <- mod_control_pa_period_view_server("period_1")
     period_buttons_2 <- mod_control_pa_period_view_server("period_2", add_period_btn = period_buttons_1$add_period_btn)
@@ -471,14 +479,15 @@ app_server <- function(input, output, session) {
   
   # Control row of inputs
     # Row 1
-    mod_report_pa_period_server("period_1")
+    mod_report_pa_period_server("period_1", dates_inputs = dates_inputs)
     mod_control_pa_period_view_server("period_1", 
                                       remove_period_btn = period_buttons_2$remove_period_btn)
     
     # Row 2
-    mod_report_pa_period_server("period_2", 
+    mod_report_pa_period_server("period_2",
                                 add_period_btn = period_buttons_1$add_period_btn, 
-                                remove_period_btn = period_buttons_2$remove_period_btn)
+                                remove_period_btn = period_buttons_2$remove_period_btn, 
+                                dates_inputs = dates_inputs)
     mod_control_pa_period_view_server("period_2", 
                                       add_period_btn = period_buttons_1$add_period_btn, 
                                       remove_period_btn = period_buttons_3$remove_period_btn)
@@ -486,7 +495,8 @@ app_server <- function(input, output, session) {
     # Row 3
     mod_report_pa_period_server("period_3", 
                                 add_period_btn = period_buttons_2$add_period_btn, 
-                                remove_period_btn = period_buttons_3$remove_period_btn)
+                                remove_period_btn = period_buttons_3$remove_period_btn, 
+                                dates_inputs = dates_inputs)
     mod_control_pa_period_view_server("period_3", 
                                       add_period_btn = period_buttons_2$add_period_btn, 
                                       remove_period_btn = period_buttons_4$remove_period_btn)
@@ -494,7 +504,8 @@ app_server <- function(input, output, session) {
     # Row 4
     mod_report_pa_period_server("period_4", 
                                 add_period_btn = period_buttons_3$add_period_btn, 
-                                remove_period_btn = period_buttons_4$remove_period_btn)
+                                remove_period_btn = period_buttons_4$remove_period_btn, 
+                                dates_inputs = dates_inputs)
     mod_control_pa_period_view_server("period_4", 
                                       add_period_btn = period_buttons_3$add_period_btn, 
                                       remove_period_btn = period_buttons_5$remove_period_btn)
@@ -502,7 +513,8 @@ app_server <- function(input, output, session) {
     # Row 5
     mod_report_pa_period_server("period_5", 
                                 add_period_btn = period_buttons_4$add_period_btn, 
-                                remove_period_btn = period_buttons_5$remove_period_btn)
+                                remove_period_btn = period_buttons_5$remove_period_btn, 
+                                dates_inputs = dates_inputs)
     mod_control_pa_period_view_server("period_5", 
                                       add_period_btn = period_buttons_4$add_period_btn, 
                                       remove_period_btn = period_buttons_6$remove_period_btn)
@@ -510,7 +522,8 @@ app_server <- function(input, output, session) {
     # Row 6
     mod_report_pa_period_server("period_6", 
                                 add_period_btn = period_buttons_5$add_period_btn, 
-                                remove_period_btn = period_buttons_6$remove_period_btn)
+                                remove_period_btn = period_buttons_6$remove_period_btn, 
+                                dates_inputs = dates_inputs)
[…]
