### 2022-11-25 91aa8c4
Added the possibility to indicate relevant missing PA periods (not yet introduced in the dataset)
