+++ R/app_server.R
@@ -452,9 +452,10 @@ app_server <- function(input, output, session) {
   #####################################################
   
   # Initializing reactive values for measurement dates
-  
   dates_inputs <- reactiveValues(dates = NULL)
   
+  # Updating reactive values for measurement dates when clicking on the "Validate
+  # configuration" button
   observeEvent(input$validate, {
     dates_inputs$dates <- attributes(as.factor(df()$date))$levels
   })
@@ -479,12 +480,15 @@ app_server <- function(input, output, session) {
   
   # Control row of inputs
     # Row 1
-    mod_report_pa_period_server("period_1", dates_inputs = dates_inputs)
+    period_info_1 <- mod_report_pa_period_server("period_1", 
+                                dates_inputs = dates_inputs)
     mod_control_pa_period_view_server("period_1", 
+                                      # Below the input is set to allow keeping visible the "Add"
+                                      # button of the first row after removing the 2nde row
                                       remove_period_btn = period_buttons_2$remove_period_btn)
     
     # Row 2
-    mod_report_pa_period_server("period_2",
+    period_info_2 <- mod_report_pa_period_server("period_2",
                                 add_period_btn = period_buttons_1$add_period_btn, 
                                 remove_period_btn = period_buttons_2$remove_period_btn, 
                                 dates_inputs = dates_inputs)
@@ -493,7 +497,7 @@ app_server <- function(input, output, session) {
                                       remove_period_btn = period_buttons_3$remove_period_btn)
     
     # Row 3
-    mod_report_pa_period_server("period_3", 
+    period_info_3 <- mod_report_pa_period_server("period_3", 
                                 add_period_btn = period_buttons_2$add_period_btn, 
                                 remove_period_btn = period_buttons_3$remove_period_btn, 
                                 dates_inputs = dates_inputs)
@@ -502,7 +506,7 @@ app_server <- function(input, output, session) {
                                       remove_period_btn = period_buttons_4$remove_period_btn)
     
     # Row 4
-    mod_report_pa_period_server("period_4", 
+    period_info_4 <- mod_report_pa_period_server("period_4", 
                                 add_period_btn = period_buttons_3$add_period_btn, 
                                 remove_period_btn = period_buttons_4$remove_period_btn, 
                                 dates_inputs = dates_inputs)
@@ -511,7 +515,7 @@ app_server <- function(input, output, session) {
                                       remove_period_btn = period_buttons_5$remove_period_btn)
     
     # Row 5
-    mod_report_pa_period_server("period_5", 
+    period_info_5 <- mod_report_pa_period_server("period_5", 
                                 add_period_btn = period_buttons_4$add_period_btn, 
                                 remove_period_btn = period_buttons_5$remove_period_btn, 
                                 dates_inputs = dates_inputs)
@@ -520,7 +524,7 @@ app_server <- function(input, output, session) {
                                       remove_period_btn = period_buttons_6$remove_period_btn)
     
     # Row 6
-    mod_report_pa_period_server("period_6", 
+    period_info_6 <- mod_report_pa_period_server("period_6", 
                                 add_period_btn = period_buttons_5$add_period_btn, 
                                 remove_period_btn = period_buttons_6$remove_period_btn, 
[…]
