+++ R/app_server.R
@@ -214,7 +214,7 @@ app_server <- function(input, output, session) {
   # Selecting days required for analysis
     output$select_days <- renderUI({
       dates <- paste0(attributes(as.factor(df()$date))$levels, " (", weekdays(as.Date(attributes(as.factor(df()$date))$levels)), ")")
-      checkboxGroupInput("selected_days", h3("Select the days to keep for analyzis (please only select the 7 appropriate days if your analyzis is related to PROactive framework)"), dates, selected = dates, inline = TRUE)
+      checkboxGroupInput("selected_days", h3("Select the days to keep for analysis (please only select the 7 appropriate days if your analysis is related to PROactive framework)"), dates, selected = dates, inline = TRUE)
     })  
     
   
@@ -613,34 +613,41 @@ app_server <- function(input, output, session) {
                                       add_period_btn = period_buttons_14$add_period_btn)
     
    
-    # Making a dataframe with the inputs set to indicate relevant missing PA periods
-    # when clicking on the Run button
-    recap_pa_perdiods <- eventReactive(input$Run, {
-       df_1  <- get_pa_period_info(period = period_info_1)
-       df_2  <- get_pa_period_info(period = period_info_2)
-       df_3  <- get_pa_period_info(period = period_info_3)
-       df_4  <- get_pa_period_info(period = period_info_4)
-       df_5  <- get_pa_period_info(period = period_info_5)
-       df_6  <- get_pa_period_info(period = period_info_6)
-       df_7  <- get_pa_period_info(period = period_info_7)
-       df_8  <- get_pa_period_info(period = period_info_8)
-       df_9  <- get_pa_period_info(period = period_info_9)
-       df_10 <- get_pa_period_info(period = period_info_10)
-       df_11 <- get_pa_period_info(period = period_info_11)
-       df_12 <- get_pa_period_info(period = period_info_12)
-       df_13 <- get_pa_period_info(period = period_info_13)
-       df_14 <- get_pa_period_info(period = period_info_14)
-       df_15 <- get_pa_period_info(period = period_info_15)
-      
-      recap <-
-        dplyr::bind_rows(
-          df_1, df_2, df_3, df_4, df_5,
-          df_6, df_7, df_8, df_9, df_10,
-          df_11, df_12, df_13, df_14, df_15
-          ) %>%
-        dplyr::filter(date != "...")
-
-      return(recap)
+    # Making a dataframe (based on the inputs set to indicate relevant missing PA periods)
+    # when clicking on the Run button ; this dataframe will be used to replace data from the
+    # initial accelerometer dataset by the data of the present dataframe for the corresponding 
+    # epochs
+    
+        # Getting inputs
+        recap_pa_perdiods <- eventReactive(input$Run, {
+           df_1  <- get_pa_period_info(period = period_info_1)
+           df_2  <- get_pa_period_info(period = period_info_2)
+           df_3  <- get_pa_period_info(period = period_info_3)
+           df_4  <- get_pa_period_info(period = period_info_4)
+           df_5  <- get_pa_period_info(period = period_info_5)
+           df_6  <- get_pa_period_info(period = period_info_6)
+           df_7  <- get_pa_period_info(period = period_info_7)
+           df_8  <- get_pa_period_info(period = period_info_8)
+           df_9  <- get_pa_period_info(period = period_info_9)
+           df_10 <- get_pa_period_info(period = period_info_10)
+           df_11 <- get_pa_period_info(period = period_info_11)
+           df_12 <- get_pa_period_info(period = period_info_12)
+           df_13 <- get_pa_period_info(period = period_info_13)
+           df_14 <- get_pa_period_info(period = period_info_14)
+           df_15 <- get_pa_period_info(period = period_info_15)
+          
+        # Making dataframe
+        recap <-
+          dplyr::bind_rows(
[…]
