### 2022-11-27 94ff105
updated figure functions: replaced `size` arguments by `linewidth` arguments where appropriate (cf. maj ggplot2 package); added alert message for the figure with steps when the score is beyond the upper limit of the x axis
