+++ R/create_fig_steps.R
@@ -59,21 +59,21 @@ language <- match.arg(language)
   band_color <- scales::seq_gradient_pal("white", "blue")(seq(0,1,length.out=6))
   
 # Creating figure 
-  if (language == "en") {
+  if (language == "en" && score <= 18000) {
     
   g_steps <-
     ggplot(data = table_steps1, aes(x = x, y = start)) +
     geom_rect(aes(ymin = start, ymax = end, fill = label), xmin = -Inf, xmax = Inf, alpha = 0.5) +
-    geom_vline(aes(xintercept = 0.40), size = 0.4, color = "grey30") +
-    geom_vline(aes(xintercept = 0.20,), size = 0.5, color = "grey30", linetype ="dotted") +
+    geom_vline(aes(xintercept = 0.40), linewidth = 0.4, color = "grey30") +
+    geom_vline(aes(xintercept = 0.20,), linewidth = 0.5, color = "grey30", linetype ="dotted") +
     geom_segment(x = 0, xend = 0.4, y = score, yend = score, 
-                 size = 0.5, color = "grey30", linetype ="dotted")   +
+                 linewidth = 0.5, color = "grey30", linetype ="dotted")   +
     geom_point(aes(x = 0.2, y = score), color = "red", size = 7, shape = 1) +
     geom_point(aes(x = 0.2, y = score), color = "red", size = 4, shape = 16) +
     geom_point(aes(x = 0.2, y = score), color = "red", size = 7, shape = 3) +
-    geom_segment(data = table_steps2, aes(x = x, xend = xend, y = y, yend = yend, color = label), size = 1, 
+    geom_segment(data = table_steps2, aes(x = x, xend = xend, y = y, yend = yend, color = label), linewidth = 1, 
                  arrow = arrow(length = unit(0.09, "inches"))) +
-    geom_errorbarh(data = table_steps2, aes(y = y, xmin = xmin, xmax = xmax, color = label), size = 1, height = 0) + 
+    geom_errorbarh(data = table_steps2, aes(y = y, xmin = xmin, xmax = xmax, color = label), linewidth = 1, height = 0) + 
     scale_y_continuous(limits = c(0, 18000),
                        labels = as.character(c(0, 2500, 5000, 7500, 10000, 12500, 18000)), 
                        breaks = c(0, 2500, 5000, 7500, 10000, 12500, 18000)) +
@@ -107,21 +107,21 @@ language <- match.arg(language)
   
   }
   
-  if (language == "fr") {
+  if (language == "fr" && score <= 18000) {
     
     g_steps <-
       ggplot(data = table_steps1, aes(x = x, y = start)) +
       geom_rect(aes(ymin = start, ymax = end, fill = label), xmin = -Inf, xmax = Inf, alpha = 0.5) +
-      geom_vline(aes(xintercept = 0.40), size = 0.4, color = "grey30") +
-      geom_vline(aes(xintercept = 0.20,), size = 0.5, color = "grey30", linetype ="dotted") +
+      geom_vline(aes(xintercept = 0.40), linewidth = 0.4, color = "grey30") +
+      geom_vline(aes(xintercept = 0.20,), linewidth = 0.5, color = "grey30", linetype ="dotted") +
       geom_segment(x = 0, xend = 0.4, y = score, yend = score, 
-                   size = 0.5, color = "grey30", linetype ="dotted")   +
+                   linewidth = 0.5, color = "grey30", linetype ="dotted")   +
       geom_point(aes(x = 0.2, y = score), color = "red", size = 7, shape = 1) +
       geom_point(aes(x = 0.2, y = score), color = "red", size = 4, shape = 16) +
       geom_point(aes(x = 0.2, y = score), color = "red", size = 7, shape = 3) +
[…]
