+++ inst/guide/user_guide_fr.Rmd
@@ -103,7 +103,11 @@ Si le patient considère que son sexe est indéfini ("undefined" dans l'applicat
 * Les catégories SED, LPA, MPA, VPA en s'appuyant sur l'axe et les valeurs seuils configurés par l'utilisateur ;
 * Les METs, en utilisant l'équation de MET choisie par l'utilisateur (si le patient considère que son sexe est indéfini ("undefined" dans l'application) ou choisi l'option "prefer not to say", alors les équations incluant une information sur le sexe, lorsque sélectionnées, sont utilisées comme si le patient était une personne de sexe féminin ; lorsque la catégorie "intersex" de l'application est utilisée, une moyenne des METs correspondant respectivement à une personne de sexe masculin et une personne de sexe féminin est utilisée avec les équations utilisant le sexe ; à noter qu'au moment de l'écriture de ce guide, il n'y a pas de donnée scientifique pour justifier un calcul pour des personnes intersexes) ;
 * Les kilocalories. Pour les périodes non-SED, les valeurs de MET calculées sont multipliées par le MB exprimé en kcal/min lorsque les équations de Santos-Lozano et al. [-@santos-lozanoActigraphGT3XValidation2013] sont utilisées, étant donné que dans cette étude, les METs étaient des multiples du métabolisme de repos mesuré (et non pas standard). Lorsque les équations de Sasaki et al. [-@sasakiValidationComparisonActiGraph2011] et de Freedson et al. [-@freedsonCalibrationComputerScience1998] sont utilisées, les valeurs de MET sont multipliées par le poids et 1/60, étant donné que dans ces études, les METs étaient des multiples du métabolisme de repos standard (i.e., 3.5 mLO$_{2}$/min/kg), et qu'un MET standard est approximativement équivalent à 1 kcal/kg/h [@butteAssessingPhysicalActivity2012]. Pour les périodes SED, le MB en kcal/min est directement utilisé ;
-* Les MET-heures liés à MVPA, en multipliant la valeur de MET par le temps (1/60e d'heure), cela seulement lorsque la valeur de MET est $\geqslant$ 3.
+* Les MET-heures liés à MVPA, en multipliant la valeur de MET par le temps, cela seulement lorsque la valeur de MET est $\geqslant$ 3.
+
+À noter que les kilocalories et les MET-heures sont initiallement calculés sur une base de 1 minute et sont ensuite ajustés en utilisant un facteur de correction pour correspondre à la durée de la période de temps choisie pour analyser les données de l'accéléromètre.
+
[…]
