+++ inst/guide/user_guide_en.Rmd
@@ -98,7 +98,11 @@ If the patient considers their sex as "undefined" or chooses the "prefer not to 
 * SED, LPA, MPA, VPA categories based on the axis and the cut-points configured by the user;
 * METs, by using the MET equation provided by the user (if the patient considers their sex as "undefined" or chooses the "prefer not to say" option, then equations including sex information, when selected, are used as if the patient were a female; when the "intersex" category is used, an average of the METs related respectively to a male and to a female is used with the equations using sex information; of note, at the time of writing this guide, there is no scientific data to justify any calculation for intersex people);
 * Kilocalories. For non-SED epochs, the computed MET values are multiplied by BMR expressed in kcal/min when using the Santos-Lozano et al. [-@santos-lozanoActigraphGT3XValidation2013] equations since, in that study, METs were multiples of the measured (not standard) resting metabolic rate. When using the  Sasaki et al. [-@sasakiValidationComparisonActiGraph2011] and Freedson et al. [-@freedsonCalibrationComputerScience1998] equations, the computed MET values are multiplied by weight and 1/60 since, in those studies, METs were multiples of standard resting metabolic rate (i.e., 3.5 mLO$_{2}$/min/kg) and a standard MET is approximately equivalent to 1 kcal/kg/h [@butteAssessingPhysicalActivity2012]. For SED epochs, BMR expressed in kcal/min is directly used;
-* MET-hours related to MVPA, by multiplying the computed MET value by the time (1/60e of an hour), only when the MET value is $\geqslant$ 3.
+* MET-hours related to MVPA, by multiplying the computed MET value by the time, only when the MET value is $\geqslant$ 3.
+
+Of note, kilocalories and MET-hours are initially computed on a 1-min basis, and are then adjusted using a correction factor to correspond to the epoch duration chosen to analyse the accelerometer dataset.
+
+The user also has the possibility to provide information about potential relevant physical activity periods that would have modified physical activity level but that could not be recorded due to accelerometer removal, as for example during some team sport or swimming activities. The user has to select the date at which the physical activity event occured. Then, two boxes have to be filled to indicate the start time of the period: one box for the number dedicated to the hour, and one box dedicated to number of the minute. Same thing for the end time of the period. Finally, the user has to indicate a MET value that would reflect the mean intensity of physical activity during the period. To do this, the user should use the [Compendium of physical activities](https://sites.google.com/site/compendiumofphysicalactivities/home?pli=1) (https://sites.google.com/site/compendiumofphysicalactivities/home?pli=1) that proposes a MET value for a wide range of activities. This information provided by the user will be used to replace data from the initial accelerometer dataset. Each line with a date that is selected will be used. If a period overlaps with a previous period, that data of the last period will be used. Importantly, using this information will modify the true wear time related to the accelerometer: the time related to self-reported physical activity will be considered as "wear" for epochs that would have been initially detected as "nonwear" when using the dedicated algorithms.
 
 Once these new variables added to the initial dataset, the app summarizes the results by day using valid wear time only, this for the following metrics: 
 
