### 2022-11-27 52c3313
updated function description for `mark_intensity()`: more precision about the fact that kcal and mets-hr are adjusted for the timestamp of the data file
