+++ tests/testthat/test-golem-recommended.R
@@ -1,7 +1,7 @@
 ##############
 # TESTS FOR UI
 ##############
-
+ 
 test_that("app ui", {
   ui <- app_ui()
   golem::expect_shinytaglist(ui)
@@ -27,7 +27,7 @@ test_that("app server", {
 })
 
 # Configure this test to fit your need
-#test_that(
+# test_that(
 #  "app launches",{
 #    golem::expect_running(sleep = 5)
 #  }
@@ -38,357 +38,323 @@ test_that("The server functions correctly work", {
 
   skip_on_cran() # This is set because PhantomJS is not found on Debian platform during CRAN submission pre-tests.
   
-  # Building a new environment with variables having to be seen globally for passing tests
-    test_activAnalyzer_env <- new.env()
-    test_activAnalyzer_env$Study <- "Study"
-    test_activAnalyzer_env$Metric <- "Metric"
-    test_activAnalyzer_env$Score <- "Score"
-    test_activAnalyzer_env$Date <- "Date"
-    test_activAnalyzer_env$wear_time <- "wear_time"
-    test_activAnalyzer_env$validity <- "validity"
-    test_activAnalyzer_env$col_time_stamp <- "col_time_stamp"
-    test_activAnalyzer_env$Range <- "Range"
-    test_activAnalyzer_env$Intervalle <- "Intervalle"
-    test_activAnalyzer_env$Selected_Day_ID <- "Selected_Day_ID"
-    test_activAnalyzer_env$Steps_score <- "Steps_score"
-    test_activAnalyzer_env$VMU_score <- "VMU_score"
-    test_activAnalyzer_env$Validity <- "Validity"
-    test_activAnalyzer_env$Item <- "Item"
-    test_activAnalyzer_env$Day <- "Day"
-    test_activAnalyzer_env$score_type <- "score_type"
-    test_activAnalyzer_env$difficulty_score_raw <- "difficulty_score_raw"
-    test_activAnalyzer_env$difficulty_score_raw_diff <- "difficulty_score_raw_diff"
-    test_activAnalyzer_env$difficulty_score_raw_quant <- "difficulty_score_raw_quant"
-    test_activAnalyzer_env$difficulty_score_rasch <- "difficulty_score_rasch"
-    test_activAnalyzer_env$amount_score_raw <- "amount_score_raw"
-    test_activAnalyzer_env$amount_score_raw_diff <- "amount_score_raw_diff"
-    test_activAnalyzer_env$amount_score_raw_quant <- "amount_score_raw_quant"
-    test_activAnalyzer_env$amount_score_rasch <- "amount_score_rasch"
-    test_activAnalyzer_env$total_score_raw <- "total_score_raw"
-    test_activAnalyzer_env$total_score_rasch <- "total_score_rasch"
-    test_activAnalyzer_env$Jour <- "Jour"
-    test_activAnalyzer_env$init <- "init"
-    test_activAnalyzer_env$users <- "users"
-    assign("equations_mets", activAnalyzer:::equations_mets, envir = test_activAnalyzer_env)
-    assign("mvpa_cutpoints", activAnalyzer:::mvpa_cutpoints, envir = test_activAnalyzer_env)
-    assign("sed_cutpoints", activAnalyzer:::sed_cutpoints, envir = test_activAnalyzer_env)
-    
+#===============================================================================
+# Setting environment
+#===============================================================================
 
-  # Attaching new environment so that variables defined above as global can be located during testing
-    attach(test_activAnalyzer_env)
+# Building a new environment with variables having to be seen globally for passing tests
+test_activAnalyzer_env <- new.env()
[…]
