### 2022-11-27 1bd96e4
maj functions doc
