### 2022-11-27 5fd9540
removed mac option from rcmd check (too instable when testing it on GitHub)
