+++ .github/workflows/R-CMD-check.yaml
@@ -18,10 +18,8 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macOS-latest,   r: 'release'}
           - {os: windows-latest, r: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
-          - {os: macOS-latest,   r: 'devel'}
           - {os: windows-latest, r: 'devel'}
           - {os: ubuntu-latest, r: 'devel'}
 
