### 2022-11-27 1be3fa0
Corrected a bug: the app crashed when changing age or weight inputs after a data file was loaded. This is now fixed.
