+++ tests/testthat/test-golem-recommended.R
@@ -103,7 +103,9 @@ app <- shinytest::ShinyDriver$new(
 #===============================================================================
 
 app$uploadFile(upload = "acc.agd")
+# app$uploadFile(upload = "inst/extdata/acc.agd")
 test_file <- "acc.agd"
+# test_file <- "inst/extdata/acc.agd"
 
 #===============================================================================
 # Testing auto-filling patient information
@@ -1865,7 +1867,306 @@ expect_equal(actual_bmr, test_bmr)
 #=============================================================================
 # Testing results obtained when setting additionnal PA periods (self-reported)
 #=============================================================================
-          
+
+# Table of inputs should initially have 0 line                        
+recap_pa_perdiods <- app$getAllValues()$export[["recap_pa_perdiods"]]
+n <- nrow(recap_pa_perdiods)                
+expect_equal(n, 0)
+
+# Setting reference inputs
+test_bmr <- 9.74 * 78 + 694
+
+ref_inputs_tab <- 
+  data.frame(
+    date = c(rep("2021-04-08", 4), rep("2021-04-09", 4), rep("2021-04-10", 4), rep("2021-04-11", 3)),
+    start_hh_num = c(rep(c(0, 1, 2, 3), 3), c(0, 1, 2)),
+    start_mm_num = rep(31, 15),
+    end_hh_num = c(rep(c(1, 2, 3, 4), 3), c(1, 2, 3)),
+    end_mm_num = rep(30, 15),
+    METS = c(rep(1, 4), rep(2, 4), rep(3, 4), rep(8, 3))
+  ) %>%
+  dplyr::mutate(
+    start = hms::as_hms(start_hh_num*3600 + start_mm_num*60),
+    end = hms::as_hms(end_hh_num*3600 + end_mm_num*60)
+  ) %>%
+  dplyr::select(date, start, end, METS)
+
+test_results_summary <- 
+  test_df %>%
+  mark_intensity(
+    col_axis = "axis1", 
+    sed_cutpoint = 100, 
+    mpa_cutpoint = 1952, 
+    vpa_cutpoint = 5725,
+    equation = "Freedson et al. (1998) [Adults]", 
+    age = 47, 
+    weight = 78, 
+    sex = "female"
+  ) %>%
+  recap_by_day(
+    age = 47, 
+    weight = 78, 
+    sex = "female"
+    ) %>%
+  dplyr::mutate(
+    wear_time = dplyr::case_when(
+      date == "2021-04-07" ~ wear_time,
+      date == "2021-04-08" ~ wear_time + 4*60,
+      date == "2021-04-09" ~ wear_time + 4*60,
+      date == "2021-04-10" ~ wear_time + 4*60,
+      date == "2021-04-11" ~ wear_time + 3*60,
+      date == "2021-04-12" ~ wear_time
+    ),
+    total_kcal = dplyr::case_when(
+      date == "2021-04-07" ~ total_kcal,
+      date == "2021-04-08" ~ total_kcal,
+      date == "2021-04-09" ~ total_kcal - 4*60*(test_bmr/24/60) + 4*60*2*(test_bmr/24/60),
+      date == "2021-04-10" ~ total_kcal - 4*60*(test_bmr/24/60) + 4*60*3*(test_bmr/24/60),
+      date == "2021-04-11" ~ total_kcal - 3*60*(test_bmr/24/60) + 3*60*8*(test_bmr/24/60),
[…]
