### 2022-11-27 #4741 @pydemull
Thank you very much @samhforbes for checking the app and the paper. I have made the corrections following your remarks. I have also performed a new read of the paper without detecting new errors but, to be honest, as a non-native english speacker, it is possible that I may not be able to manage all the the subtleties of the English language if any.
