### 2022-11-28 #4741 @samhforbes
Hi @pydemull great that looks like it's cleaned it up.
Can you please make a tagged release, and archive that somewhere with a DOI (such as Zenodo for instance) and then post the DOI here?
