### 2022-11-28 #4741 @pydemull
Sorry @samhforbes I am not sure of the materials I have to put on Zenodo. Is it the source code of the version 1.0.5 ? (this version does not contain all the updates made after the comments from one of the reviewer). Or is it the source code of the development version that is currently on the master branch? (knowing that some other improvements have been made since the feedbacks from the reviewers).
