### 2022-11-29 a673a08
maj workflow and tests
