### 2022-11-29 7664b61
maj test mark_wear_time function
