### 2022-11-30 #4741 @samhforbes
Hi @pydemull sorry I missed this coming through. You want a tagged release of the post review version, which is what gets put on Zenodo.
