### 2022-11-30 #4741 @pydemull
Ok thanks a lot @samhforbes. So, I will tagg a version 1.1.0, because the current version contains ,in addition to the features asked by the reviewers, two new features: possibility to zoom in on the graphics of the app, and the possibility to manually add physical activity information (eg, relating to what has been done during device removal periods) into the accelerometer dataset. I will update the article using this version number.
