+++ joss/paper.md
@@ -40,7 +40,7 @@ Due to the large size of data files to analyse when using accelerometry and due 
 Other ways than ActiLife software to analyse activity counts include using programming languages. R [@rcoreteamLanguageEnvironmentStatistical2022] and Python [@pythonsoftwarefoundationPython2022] have been programming languages commonly used by scientists to build tools aiming at fostering physical activity data analysis. In R, the 'accelerometry' and 'nhanesaccel' packages by Van Domelen and Pittard [-@vandomelenFlexibleFunctionsProcessing2014], the 'actigraph.sleepr' package by Petkova [-@petkovaActigraphSleeprDetect2021], and the 'pawacc' package by Geraci [-@geraciPawaccPhysicalActivity2017], provide several functions to perform analyses of interest with activity counts. In Python, the 'pyActigraphy' library by Hammad and Reyt [-@hammadPyActigraphy2020] also allows, among various other features, to handle ActiGraph activity counts. While useful for research settings, these ressources may be of a little interest for other settings where people have no programming skills, because they do not propose a GUI (graphical user interface) to help people who do not code and who have no time to learn this skill. Beyond the lack of a free and simple interface to analyse ActiGraph activity counts data, there is, to our knowledge, no app that allows an easy implementation of the PROactive framework with COPD patients that would be based on an analysis of ActiGraph activity counts. This is why we have developed the 'activAnalyzer' app. For now, a first main interest of this app is to allow teaching large groups of students or professionnals, who have no programming skills, to analyse activity counts for assessing physical behaviour. A second main interest is to allow an easy implementation of the PROactive framework with COPD patients when working with an ActiGraph accelerometer, this by clinicians, healthcare providers and/or researchers, either in clinical routine or in research setting.
 
 #  Functionality and purpose of the software
-'activAnalyzer' is an app built as a package using R programming language. The app can be used in three different ways as explained elsewhere (https://pydemull.github.io/activAnalyzer/), including (i) a standalone desktop application for Windows machines only thanks to the [DesktopDeployR framework developed by Lee Pang](https://github.com/wleepang/DesktopDeployR); (ii) using [R](https://CRAN.R-project.org/) and [RStudio](https://www.rstudio.com/) software along with the CRAN version of the 'activAnalyzer' package (v1.0.5) or its development version from GitHub; (iii) using a [shinyapps.io plateform ](https://pydemull.shinyapps.io/activAnalyzer/). If used with R and RStudio, the app will require the installation of the [TinyTeX distribution](https://yihui.org/tinytex/) to generate .pdf reports, as explained on the [app website](https://pydemull.github.io/activAnalyzer/).
+'activAnalyzer' is an app built as a package using R programming language. The app can be used in three different ways as explained elsewhere (https://pydemull.github.io/activAnalyzer/), including (i) a standalone desktop application for Windows machines only thanks to the [DesktopDeployR framework developed by Lee Pang](https://github.com/wleepang/DesktopDeployR); (ii) using [R](https://CRAN.R-project.org/) and [RStudio](https://www.rstudio.com/) software along with the CRAN version of the 'activAnalyzer' package or its development version from GitHub; (iii) using a [shinyapps.io plateform ](https://pydemull.shinyapps.io/activAnalyzer/). If used with R and RStudio, the app will require the installation of the [TinyTeX distribution](https://yihui.org/tinytex/) to generate .pdf reports, as explained on the [app website](https://pydemull.github.io/activAnalyzer/).
 
[…]
