### 2022-11-30 #4741 @pydemull
@samhforbes please find here the DOI: 10.5281/zenodo.7384191
Regarding the article, I have removed from the text the version number associated to the CRAN version for more simplicity. The current version is thus 1.1.0.
Best
