### 2022-12-01 #4741 @Kevin-Mattheus-Moerman
@pydemull I am the AEiC for this track and here to help process this work for acceptance in JOSS. I have reviewed the archive and the paper and have noted the below issues that require your attention:

On the archived software link https://zenodo.org/record/7384191:
- [ ] Please manually edit the title to fully match that of the paper
- [ ] Also edit the author set to be the same, and in the same order, as the paper.
- [ ] Also edit the license to match your software license.
(@samhforbes thanks for editing this work! As a minor point of feedback, check for these points in the future as well ☝️)

- [ ] Can you change software names like ‘accelerometry’ to use ` symbols, instead of ‘. E.g.
`accelerometry`
```
This way they are highlighted properly.
- [ ] Change `ressources` to `resources`
- [ ] Check `... activity and sedentary behaviours remains common`, it looks like you should use `behaviour` (singular form).
- [ ] Fix `professionnals` to `professionals`
- [ ] Fix `plateform` to `platform`
- [ ] Recommendation: Consider rephrasing `...are only from the authors of the present work.` to `...rested with the authors of the present work.`, or feel free to propose an alternative.
- [ ] Fix `feedbacks` to `feedback`
- [ ] Can you confirm the city/town names are included in the affiliations? Can you expand/define the acronym GHICL?

Let me know when you've addressed the above. Thanks.
