+++ joss/paper.md
@@ -25,22 +25,21 @@ authors:
   affiliation: 2
 bibliography: references.bib
 affiliations:
-- name: Institute of Physical Education and Sport Sciences (IFEPSA-UCO), Les Ponts-de-Cé,
-    France
+- name: Institut de formation en éducation physique et en sport d'Angers (IFEPSA-UCO), Les Ponts-de-Cé, France
   index: 1
-- name: Hôpital Saint Philibert, GHICL, France
+- name: Groupement des Hôpitaux de l’Institut Catholique de Lille (GHICL), Hôpital Saint Philibert, Lomme, France
   index: 2
 ---
 
 # Statement of need
-ActiGraph devices (ActiGraph LLC, Pensacola, FL) have been the most used accelerometers in scientific literature to measure physical activity [@bassettAccelerometerbasedPhysicalActivity2015a; @miguelesAccelerometerDataCollection2017]. These devices, along with their software companion ActiLife (ActiGraph LLC, Pensacola, FL), allow to get movement data expressed in either *activity counts* or *G-force* units. Activity counts represent the amount of acceleration produced over a given epoch of time at the wearing position of the device. While analytic methods based on G-force data are developing, using activity counts to assess physical activity and sedentary behaviours remains common [@miguelesAccelerometerDataCollection2017]. Interestingly, ActiGraph devices have been validated to implement the framework related to the use of the PROactive Physical Activity in COPD (chronic obstructive pulmonary disease) instruments [@dobbelsPROactiveInnovativeConceptual2014; @gimeno-santosPROactiveInstrumentsMeasure2015; @garcia-aymerichValidityResponsivenessDaily2021]. Such a framework requires combining scores relating to answers of questionnaire items and scores relating to accelerometer metrics obtained from a week of measurement.
+ActiGraph devices (ActiGraph LLC, Pensacola, FL) have been the most used accelerometers in scientific literature to measure physical activity [@bassettAccelerometerbasedPhysicalActivity2015a; @miguelesAccelerometerDataCollection2017]. These devices, along with their software companion ActiLife (ActiGraph LLC, Pensacola, FL), allow to get movement data expressed in either *activity counts* or *G-force* units. Activity counts represent the amount of acceleration produced over a given epoch of time at the wearing position of the device. While analytic methods based on G-force data are developing, using activity counts to assess physical activity and sedentary behaviour remains common [@miguelesAccelerometerDataCollection2017]. Interestingly, ActiGraph devices have been validated to implement the framework related to the use of the PROactive Physical Activity in COPD (chronic obstructive pulmonary disease) instruments [@dobbelsPROactiveInnovativeConceptual2014; @gimeno-santosPROactiveInstrumentsMeasure2015; @garcia-aymerichValidityResponsivenessDaily2021]. Such a framework requires combining scores relating to answers of questionnaire items and scores relating to accelerometer metrics obtained from a week of measurement.
 
-Due to the large size of data files to analyse when using accelerometry and due to the relative complexity of the implementation of some algorithms, a simple spreadsheet does not appear to be a feasible tool to complete all the accelerometer data analysis workflow from the same place. The 'Full' version of ActiLife software allows completing all the general steps of the data analysis workflow with activity counts but the cost of this ActiLife version may prevent teaching a wide audience to implement such a workflow and may also prevent working with large teams on recorded data ('Lite' versions of ActiLife software, that are at a lower cost than 'Full' versions, allow device initialisation and data downloading only). Moreover, there is no solution in ActiLife software to fully implement the PROactive framework for COPD patients described above.
[…]
