### 2022-12-01 #4741 @pydemull
Thanks (again!) @samhforbes. Great to have JOSS and the opportunity to publish software-related work. @Kevin-Mattheus-Moerman, I have made the corrections you required (thanks for this). The new version of the paper has been uploaded on the master branch.
