### 2022-12-01 #4741 @Kevin-Mattheus-Moerman
@pydemull congratulations on this publication in JOSS!

@samhforbes Thanks for editing this submission!

Special thanks also to the reviewers: @elimillera and @angerhang
